renamed support to shared
