+++ src/features/assessments/components/AllNotes.tsx
@@ -1,4 +1,4 @@
-import { getImagePath } from '@/support/helpers/ImageHelper';
+import { getImagePath } from '@/shared/helpers/ImageHelper';
 import { withActivePetAssessments } from '@core/database/hoc';
 import { Assessment } from '@core/database/models/Assessment';
 import { Pet } from '@core/database/models/Pet';
