+++ src/features/onboarding/screens/WelcomeScreen.tsx
@@ -1,5 +1,5 @@
 import { WelcomeScreenNavigationProps } from '@/features/navigation/types.tsx';
-import { CustomPagination } from '@/support/components/CustomSwiperPagination';
+import { CustomPagination } from '@/shared/components/CustomSwiperPagination';
 import React, { useEffect, useMemo, useRef, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { Animated, Dimensions, Platform, SafeAreaView, ScrollView, StyleSheet, View } from 'react-native';
