+++ src/shared/components/PulsatingCircle.tsx
