+++ src/shared/components/RatingSliderToolTip.tsx
